multi_agent_travel_planner/requirements.txt
# Core Dependencies
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
python-dotenv>=1.0.0

# Search Tools
tavily-python>=0.3.0
langchain-tavily>=0.0.1
google-search-results>=2.4.2

# Utilities
ipython>=8.12.0
